Graph Indexes

Indexes

General Graphs

GraphIdx.NeighborIndex — Type.

NeighborIndex(n, head, tail)

Construct an index for n nodes of the (undirected) neighbors given by the edges zip(head, tail). If head[i] <= 0 then edge i is excluded.

source

NeighborIndex(n, edges)

Same but for edges like [(1, 2), (2, 3)]

source

NeighborIndex(n, m, iter::Function)

Provide an iterator over the edges

source

Graph Partition

GraphIdx.Cluster.cluster — Method.

cluster(x, neigh, eps=1e-5, seed=42)

Find a partition of x such that abs(x[i] - x[j]) < eps for all edges j in neighidx[i] where i and j are in the same partition.

source

Tree Graphs

GraphIdx.Tree.ChildrenIndex — Type.

Provide constant time access to children of a node by index operator. By convention, the root node is stored at ChildrenIndex.idx[1].

source

Once, memory has been allocated we can re-use it to store different tree (of same size!) in it

GraphIdx.Tree.reset! — Method.

reset!(cidx, parent [,root])

Actually compute the index according to parent.

source