Tree Graphs

Tree Graphs

GraphIdx.Tree — Module.

Bundle functions regarding tree graphs.

source

For computing a spanning tree from within a general graph, see here.

Data Structures

We provide some framework specialized for (rooted) tree graphs:

GraphIdx.Tree.RootedTree — Type.

A rooted tree. Often you might want to have a ChildrenIndex.

source

GraphIdx.Tree.find_root — Function.

find_root(pi)

Find the first index i with pi[i] == i. Report an error if non such exist.

source

GraphIdx.Tree.node_degrees — Function.

node_degrees(parent, [root = 1])

Return a vector telling the degree of each node.

source

Often, we will need to access the children of arbitrary nodes; for this the GraphIdx.Tree.ChildrenIndex was developed.

Traversal

Classical depth first search

GraphIdx.Tree.dfs_walk — Function.

dfs_walk(f, tree [,stack])

Call f on each node of tree in depth-first search (DFS) order. Hereby the node v is negative v < 0 if the node is discovered the first time and non-negative (v >= 0) if the node has been finished.

The tree can be given as ChildrenIndex or parent Vector{Int}; in the latter case ChildrenIndex will be constructed. To avoid allocation, you can pass a stack Vector.

source

For binary trees, there is also the inorder:

GraphIdx.Tree.binary_inorder — Function.

Return the Inorder traversal of a complete binary tree with given height. (Root node will be 0 and placed in the middle of the array

source

Lowest Common Ancestors

GraphIdx.Tree.lowest_common_ancestors — Function.

lowest_common_ancestors(tree, parent, pairs)

For a list of pairs [(u1, v1), ..., (uk, vk)] compute the lowest common ancestors [a1, ..., ak], i.e. ai is the lowest node in the tree that is ancestor (parent, grand-parent, grand-grand-parent, etc.) of both, ui and vi.

Runtime is O(length(tree) + length(pairs)). More precisely, the inverse Ackermann function is at most 4 for all practical instances.

See CLRS (3rd edition) page 584 or networkx [n].

Compare to [n]: networkx/algorithms/lowestcommonancestors.py [1]: https://en.wikipedia.org/wiki/Lowestcommonancestor

source